+++ src/connectedk8s/HISTORY.rst
@@ -6,6 +6,8 @@ Release History
 1.9.0
 ++++++
 * Added support for arc gateway feature.
+* Bug fix for Workload Identity poll or agent state
+* Bug fix for connectedk8s update with workload identity enabled
 
 1.8.0
 ++++++